Add a Patch class
The purpose of Patch is to replace CSV lines in order, updating the
schema if the CSV file as needed.

For example, if the first changed row has columns A and B, and the
next row has columns A, B, and C, the column C should be added to the
schema and every next row should at least contain a null in that
column.