Update dependencies
These had remained untouched for years. We update to a more recent
version of Ruby and update the dependencies.

The Ruby version we're updating to is still old, but it aligns with the
version used by project for which csv_patch was created. We're very open
to bumping the Ruby version again if it becomes useful to someone else.